+++ lib/src/utils/utils.dart
@@ -5,6 +5,7 @@ import 'dart:convert';
 import 'dart:io' show Platform;
 import 'dart:math';
 import 'dart:typed_data';
+import 'dart:developer' as dev;
 
 import 'package:asn1lib/asn1lib.dart';
 import 'package:blockchain_utils/blockchain_utils.dart';
@@ -19,3 +20,4 @@ part 'abi_coder.dart';
 part 'crypto.dart';
 part 'p256.dart';
 part 'uint256.dart';
+part 'logger.dart';